+++ test/ibmq/test_proxies.py
@@ -25,10 +25,14 @@ from requests.exceptions import ProxyError
 from qiskit.providers.ibmq.api_v2.clients import (AuthClient,
                                                   VersionClient)
 from qiskit.providers.ibmq.api_v2.exceptions import RequestsApiError
-
 from ..ibmqtestcase import IBMQTestCase
 from ..decorators import requires_qe_access, requires_new_api_auth
 
+# Fallback mechanism. Version variable is stored under __doc__ in new pproxy versions
+try:
+    from pproxy.__doc__ import __version__ as pproxy_version
+except ImportError:
+    from pproxy import __version__ as pproxy_version
 
 ADDRESS = '127.0.0.1'
 PORT = 8080
@@ -37,15 +41,17 @@ INVALID_PORT_PROXIES = {'https': '{}:{}'.format(ADDRESS, '6666')}
 INVALID_ADDRESS_PROXIES = {'https': '{}:{}'.format('invalid', PORT)}
 
 
-@skipIf(sys.version_info >= (3, 7), 'pproxy version not supported in 3.7')
+@skipIf((sys.version_info > (3, 5) and pproxy_version == '1.2.2') or
+        (sys.version_info == (3, 5) and pproxy_version > '1.2.2'),
+        'pproxy version is not supported')
 class TestProxies(IBMQTestCase):
     """Tests for proxy capabilities."""
 
     def setUp(self):
         super().setUp()
-
+        listen_flag = '-l' if pproxy_version >= '1.7.2' else '-i'
         # launch a mock server.
-        command = ['pproxy', '-v', '-i', 'http://{}:{}'.format(ADDRESS, PORT)]
+        command = ['pproxy', '-v', listen_flag, 'http://{}:{}'.format(ADDRESS, PORT)]
         self.proxy_process = subprocess.Popen(command, stdout=subprocess.PIPE)
 
     def tearDown(self):
@@ -135,4 +141,4 @@ def pproxy_desired_access_log_line(url):
     """Return a desired pproxy log entry given a url."""
     qe_url_parts = urllib.parse.urlparse(url)
     protocol_port = '443' if qe_url_parts.scheme == 'https' else '80'
-    return 'http {}:{}'.format(qe_url_parts.hostname, protocol_port)
+    return '{}:{}'.format(qe_url_parts.hostname, protocol_port)