+++ CHANGELOG.md
@@ -18,7 +18,7 @@ The format is based on [Keep a Changelog].
 
 ### Added
 
-- Added support for backend autocompletion. A user can now use 
+- Added support for autocompleting backend names. A user can now use 
 `AccountProvider.provider_backends.<tab>` to see a list of backend names (\#303).
 
 ### Changed